% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{fdr_efdr_boxplot}
\alias{fdr_efdr_boxplot}
\title{True fdr versus estimated fdr}
\usage{
fdr_efdr_boxplot(bp, facet_fdr = TRUE)
}
\arguments{
\item{bp}{a \code{boxplot_prep} object}

\item{facet_fdr}{if \code{TRUE}, use \code{facet_wrap} across the fdr levels}
}
\value{
a \code{ggplot2} object
}
\description{
Plot the true fdr versus the estimated fdr.
Note, this requires that \code{boxplot_prep} be run ahead of time
}

