% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{fdr_efdr_plot}
\alias{fdr_efdr_plot}
\title{FDR versus estimated FDR}
\usage{
fdr_efdr_plot(de_bench)
}
\arguments{
\item{de_bench}{a \code{de_benchmark} object resulting from \code{merge_results}}
}
\value{
a \code{ggplot} object
}
\description{
Plot the true fdr versus the estimated fdr
}

