% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{fdr_efdr_power_plot}
\alias{fdr_efdr_power_plot}
\title{fdr versus power}
\usage{
fdr_efdr_power_plot(mean_fdr, fdr_levels = c(0.01, 0.05, 0.1), start = 50,
  jump = 50, rank_fdr = NULL, method_colors = NULL,
  fdr_level_position = -0.001, ignore_estimated_fdr = NULL,
  isolines = TRUE)
}
\arguments{
\item{mean_fdr}{an object from \code{get_fdr}}

\item{fdr_levels}{a vector of levels}

\item{start}{where to start the isolines}

\item{jump}{how big each step is between the isolines}

\item{rank_fdr}{if \code{TRUE}, legend based off the fdr}

\item{ignore_estimated_fdr}{if not \code{NULL}, a character vector of method names to ignore when plotting estimated fdr point}

\item{sim_filter}{if \code{TRUE}, remove things that didn't pass the simulation filter from the truth}
}
\value{
a \code{ggplot2} object
}
\description{
true fdr versus power. also point out where the estimated fdr is.
also draws isolines specifying the number of things in the ranking
}

