% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{fdr_nde_plot}
\alias{fdr_nde_plot}
\title{FDR versus ranked list}
\usage{
fdr_nde_plot(de_bench, estimate = TRUE)
}
\arguments{
\item{de_bench}{a \code{de_benchmark} object resulting from \code{merge_results}}

\item{estimate}{if \code{TRUE}, then plot the estimated FDR with dotted lines}
}
\value{
a \code{ggplot} object
}
\description{
Plot the true FDR versus the ranking induced by p-value
}

