% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{fdr_power_plot}
\alias{fdr_power_plot}
\title{true fdr versus power}
\usage{
fdr_power_plot(list_bench, sim_filter = FALSE)
}
\arguments{
\item{list_bench}{a list of \code{de_benchmark}s}
}
\value{
a ggplot2 object
}
\description{
plot the true fdr versus power
}

