% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{fdr_tpr_plot}
\alias{fdr_tpr_plot}
\title{Plot the (estimated) FDR against the TPR}
\usage{
fdr_tpr_plot(de_bench)
}
\arguments{
\item{de_bench}{a \code{de_benchmark} object}
}
\value{
a \code{ggplot} object
}
\description{
Plot the (estimated) FDR against the TPR.
Currently can only do the estimated FDR.
Will eventually support the true FDR
}

