% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{filter_benchmark}
\alias{filter_benchmark}
\title{Filter a benchmark}
\usage{
filter_benchmark(de_bench, to_remove, join_mode, keep_colors = FALSE)
}
\arguments{
\item{de_bench}{a 'de_benchmark' object}

\item{to_remove}{a character string of labels indicating which results to remove}

\item{keep_colors}{if \code{TRUE}, keep the original colors}
}
\value{
a \code{de_benchmark} object with \code{to_remove} removed
}
\description{
Filter a benchmark
}

