% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{get_sensitivity_specificity_oracle}
\alias{get_sensitivity_specificity_oracle}
\title{Get the sensitivity and specificity at a level}
\usage{
get_sensitivity_specificity_oracle(de_bench, fdr_level = c(0.01, 0.05, 0.1))
}
\arguments{
\item{de_bench}{a \code{de_benchmark} object}

\item{fdr_level}{a vector of fdr levels (e.g. \code{c(0.01, 0.05, 0.10)})}
}
\value{
a data frame with sensitivity, specificity, precision, accuracy
}
\description{
Given some levels, calculate the (estimated) sensitivity and specificity.
Note: this requires that the oracle does NOT have a column 'qval'
}

