% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{power_efdr_boxplot}
\alias{power_efdr_boxplot}
\title{power at a specific estimated fdr}
\usage{
power_efdr_boxplot(bp, facet_fdr = TRUE)
}
\arguments{
\item{bp}{a \code{boxplot_prep} object}

\item{facet_fdr}{if \code{TRUE}, use \code{facet_wrap} across the fdr levels}
}
\value{
a \code{ggplot2} object
}
\description{
Plot the estimated fdr versus power
Note, this requires that \code{boxplot_prep} be run ahead of time
}

