% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_benchmark.R
\name{rename_benchmark}
\alias{rename_benchmark}
\title{rename a benchmark}
\usage{
rename_benchmark(de_bench, original_names, new_names, join_mode)
}
\arguments{
\item{de_bench}{a \code{de_benchmark} object}

\item{original_names}{the names of the original method to be renamed}

\item{new_names}{the new names to replace the original names}
}
\value{
a new \code{de_benchmark} object
}
\description{
rename a benchmark object
}

